Move version package in cm
