+++ .ci/build
@@ -52,7 +52,7 @@ if [[ -z "$LOCAL_BUILD" ]]; then
     -a \
     -v \
     -o "${BINARY_PATH}/rel/docforge-linux-amd64" \
-    -ldflags "-w -X github.com/gardener/docforge/pkg/version.Version=${VERSION}" \
+    -ldflags "-w -X github.com/gardener/docforge/cmd/version.Version=${VERSION}" \
     cmd/*.go
 
   echo "Building docforge ${BINARY_PATH}/rel/docforge-linux-arm64"
@@ -60,7 +60,7 @@ if [[ -z "$LOCAL_BUILD" ]]; then
     -a \
     -v \
     -o "${BINARY_PATH}/rel/docforge-linux-arm64" \
-    -ldflags "-w -X github.com/gardener/docforge/pkg/version.Version=${VERSION}" \
+    -ldflags "-w -X github.com/gardener/docforge/cmd/version.Version=${VERSION}" \
     cmd/*.go
 
   echo "Building docforge ${BINARY_PATH}/rel/docforge-darwin-amd64"
@@ -68,7 +68,7 @@ if [[ -z "$LOCAL_BUILD" ]]; then
     -a \
     -v \
     -o "${BINARY_PATH}/rel/docforge-darwin-amd64" \
-    -ldflags "-w -X github.com/gardener/docforge/pkg/version.Version=${VERSION}" \
+    -ldflags "-w -X github.com/gardener/docforge/cmd/version.Version=${VERSION}" \
     cmd/*.go
 
   echo "Building docforge ${BINARY_PATH}/rel/docforge-darwin-arm64"
@@ -76,7 +76,7 @@ if [[ -z "$LOCAL_BUILD" ]]; then
     -a \
     -v \
     -o "${BINARY_PATH}/rel/docforge-darwin-arm64" \
-    -ldflags "-w -X github.com/gardener/docforge/pkg/version.Version=${VERSION}" \
+    -ldflags "-w -X github.com/gardener/docforge/cmd/version.Version=${VERSION}" \
     cmd/*.go
 
   echo "Building docforge ${BINARY_PATH}/rel/docforge-windows-386.exe"
@@ -84,7 +84,7 @@ if [[ -z "$LOCAL_BUILD" ]]; then
     -a \
     -v \
     -o "${BINARY_PATH}/rel/docforge-windows-386.exe" \
-    -ldflags "-w -X github.com/gardener/docforge/pkg/version.Version=${VERSION}" \
+    -ldflags "-w -X github.com/gardener/docforge/cmd/version.Version=${VERSION}" \
     cmd/*.go
 
 # If the LOCAL_BUILD environment variable is set, we simply run `go build`.
@@ -92,6 +92,6 @@ else
   go build \
     -v \
     -o "${BINARY_PATH}/docforge" \
-    -ldflags "-w -X github.com/gardener/docforge/pkg/version.Version=$(git rev-parse HEAD)" \
+    -ldflags "-w -X github.com/gardener/docforge/cmd/version.Version=$(git rev-parse HEAD)" \
     cmd/*.go
 fi
