+++ cmd/version/version.go
@@ -7,7 +7,6 @@ package version
 import (
 	"fmt"
 
-	"github.com/gardener/docforge/pkg/version"
 	"github.com/spf13/cobra"
 )
 
@@ -19,7 +18,12 @@ func NewVersionCmd() *cobra.Command {
 		Use:   "version",
 		Short: "Print the version",
 		Run: func(cmd *cobra.Command, args []string) {
-			fmt.Println(version.Version)
+			fmt.Println(Version)
 		},
 	}
 }
+
+// Version is a global variable which is set during compile time via -ld-flags in the `go build` process.
+// It stores the version of the Gardener and has either the form <X> or <X.Y>, where <X> denominates
+// the current 'major' version, and <Y> (if present) denominates the current 'hotfix' version.
+var Version = "binary was not built properly"
